servo: Merge #10774 - Improve properties.mako.rs file structure, take 2 (from perlun:improve-mako-file-structure-v2); r=SimonSapin
This is a new attempt of #10586, after Simon Sapin's great cleanups in #10749 has landed. I have adjusted the changes to the new structure that was introduced, and also only done a few of the longhand ones. Will certainly continue on this as soon as we have a basic agreement that this style is reasonable.

Source-Repo: https://github.com/servo/servo
Source-Revision: 69acd8e1ccba8ff01e8f7fde4305956ba5ab866d