servo: Merge #11636 - Fix an incorect behavior, when converting optional services (from szeged:optservicesfix); r=jdm
As specified in https://webbluetoothcg.github.io/web-bluetooth/#ref-for-blacklisted-2 if any service in the optionalServices are blacklisted, there is no need to throw an error (just remove the service). In this case we don't add the service to the result of conversion.
<!-- Thank you for contributing to Servo! Please replace each `[ ]` by `[X]` when the step is complete, and replace `__` with appropriate data: -->
- [X] `./mach build -d` does not report any errors
- [X] `./mach test-tidy` does not report any errors
- [ ] These changes fix #__ (github issue number if applicable).

<!-- Either: -->
- [ ] There are tests for these changes OR
- [X] These changes do not require tests because there are no webbluetooth test api implementation yet.

<!-- Pull requests that do not address these steps are welcome, but they will require additional verification as part of the review process. -->

Source-Repo: https://github.com/servo/servo
Source-Revision: a8b3a07ea2498cb6dcddf6251c71d384153ba9a7